[Bug 809617 - Configuration editor issues with set/unset, drop down lists, and map of simples] When editing a config "unset" should disable the field and nullify the value. When setting (unchecking the "unset" checkbox) the field should be enabled and focus should be given to the field.  If navigating away without setting a value the field should automatically be unset.  Behavior should be consistent when toggling "unset" whether it is for the current field or a different field.
There have issues with this behavior going back to the beginnings of
the new gui impl.  The scenario in this BZ is a particular bad example.
Also, the upgrade to smartgwt 3.0 seems to have affected things a bit
negatively as well.

This fixes many of these issues, although given the complexity of the
component I'd be surprised if it fixed them all.

Additionally, the "Delete" SelectItem for a Map of Simples was no longer
properly displaying the values after the upgrade.  I've solved this
although it's maybe not super visually pleasing (but unless there is
a complaint I'm not going to spend any more time dorking with it).