[Bug 813917 - High CPU utilization after enabling certain Metric Collection Templates for Tomcat WARs] The problem is actually independent of plugin or resource type.  It is a general problem with enable/disable of metrics at the template level. - The database is not compromised - It occurs only when applying the changes to existing inventory - It affects only running agents that are updated as a result of the   changes. - The server code leaks bad collection interval values, used to indicate   various enable/disable scenarios, to the agent update.
This has been corrected.  The changes are limited to the server jar.
There are additions to the MeasurementScheduleRemote and Local, and
certain methods have been deprecated.  Existing CLI scripts should be
ok, and I've added better validation of intervals being set in that
way. Although, if they are using the deprecated methods they should move
to the new methods after the next upgrade.

Agents do not need to be updated. But, agents suffering from this problem
should be re-synced, or restarted --purge.