Back out changes to core-gui.war introduced in commit 19bf99967bd27f1ef6de4db93d2cad13b5a1ce65 by Bug 813917.
We do not want to introduce this change into core-gui.war as this is part of a hot-fix patch and core-gui.war is patched separately from the server components.

Patches can not require bi-directional dependencies across components. This reverts core-gui to use the deprecated MeasurementScheduleManagerLocal methods. The underlying code in the deprecated methods does the same thing as their replacements.