[BZ 822299] Display better error message and no exception when agent is unreachable
This commit introduces the class CannotConnectToAgentException which is
a wrapper around the CannotConnectException that gets thrown when a
request is sent to an agent that is down or otherwise unreachable.
Previously we just propagated the exception up the call stack. In the
UI, we were handling as a generic RuntimeException and also showing the
exception stack trace in the message center. With this commit we no
longer log the exception in the UI and instead just log a more
informative warning message. This change effects manual import, resource
create, and resource delete.