Making changes to better support denormalization in the data model
This commit is primarily focused on the data model for
MeasurementDataTrait, though the changes likely will apply to other
types of MeasurementData as well. In the resource traits view in the UI
we perform a join to fetch fields from the measurement def with
which the trait is associated. A lot of NoSQL databases do not have any
support for joins; so, this commit introduces some changes that will
prevent having to query both the metrics (presumably nosql) database as
well as the rhq database to get the data to display in the traits view.

I saw two ways to go about persisting the measurement def data along
with the trait when the server receives the measurement report. The
first way would be to load the measurement defs when persisting the
report data. I elected not to go that route because that could
potentially be a lot of reads on what is already a performance intensive
operation.

I instead decided to add the necessary data to
MeasurementScheduleRequest. When a plugin receives a schedule request to
obtain a metric value, the plugin component creates the MeasurementDataTrait
object passing the request as an argument. The request is now persisted
as a field in the trait object. When the report is sent to the server,
all the data needed for denormalization is already there which allows us
to avoid any extra database hits.