Initial commit for TraitMeasurement and updating TraitsView to use it
MeasurementDataTrait implements TraitMeasurement and TraitsView (for an
individual resource) has been refactored to use TraitMeasurement instead
of the JPA specific MeasurementDataTrait. Most of the refactoring was
done in the data source classes. As part of this prototype effort, I
want to first focus on resource level view/operations, so the
refactoring was done in such a way to leave TraitsView and
TraitsDataSource for the group UI intact. It still directly uses
MeasurementDataTrait.

This is also the initial commit for MetricsGWTService which is a wrapper
for MetricsManagerBean.