[BZ 786614] Add some error logging  if we cannot find a resource container
We should not be in state where the method getAbsoluteBaseDirectory is
passed a resource id that does not have a corresponding resource
container in the local inventory. If that does happen, we will now log
an error message that includes the resource id and drift definition to
facilitate further debugging. Then an IllegalArgumentException is
thrown.