Removing all the get*Manager() methods on the RhqFacade and replacing it with one generic getProxy(Class) method. This method returns a proxy of the original remote API interface (i.e. the interface is NOT simplified).
The getManagers() method returns a map of manager proxies where
the interfaces ARE simplified.

This is in preparation for making the simplified interfaces NOT implement
the original remote APIs so that we have greater freedom in what can be done
with the API and also make code completion significantly more clean and
simple.