[830463 - Re-enabling (actually, updating) an alert definition clears condition logs for existing alerts] This is sort of complex. The underlying issue is that AlertDefinitionManagerBean.updateAlertDefinition(updatedDef) replaces existing AlertConditions with updatedDef.conditions.  The detached conditions were deleted (via Hibernate and DELETE_ORPHAN) regardless of the fact that they very possibly linked to AlertConditionLog records for existing Alerts.  These log records are what you need to display the actual condition evaluations that explain why the alert was fired.
The fix was to allow the detached AlertCondition records to stay around and
to deal with the fallout of doing that:
- clean up orphaned AlertConditions (no def and no logs) in the data purge job
- update queries backing AlertManager.deleteAlertsByContext()
- update queries backing resource uninventory bulk delete

Additionally, there was a lot of confusion around the 'purgeInternals' parameter
for updateAlertDefinition().  It was initially thought that setting this true
was responsible for the loss of AlertConditions backing the Alerts.  Actually,
what this flag does, when set to true, is to erase any partial condition
matching that has taken place.  It must be set true if conditions, dampening rules,
or the condition expression (All vs Any matching) are being updated.  Otherwise
it can be false.

I renamed this flag to be 'resetMatching' to hopefully be more clear, and
added some jdoc for the relevant methods.

I had already refactored CoreGUI's editor to correctly set this flag before
realizing it was not the root cause.  So that logic is in place and
will optimize the update to some degree.

I think the backend logic can be further optimized using this setting more
intelligently.  I plan a subsequent check-in for that before setting to ON_QA.