Getting rid of MeasurementAggregate interface
I introduced the MeausrementAggregate interface to facilitate populating
the numeric, roll up tables for testing. There is already a class in
server/jar of the same name, and this creating some confusion. Since the
class I introduced is needed for testing, I renamed the impl to
AggregateTestData to better reflect that. I also moved it to the package
where the other metrics plugin classes are. I do not want to the
MeasurementAggregate in server/jar or classes that consume it to avoid
changes that can otherwise be avoided.