Changing signature of findDataForContext to take a MeasurementSchedule
The existing, legacy implementation takes a definition id and performs a
join to get the schedule. For the Cassandra implementation I am
working on metric data is keyed off of schedule id. By changing
findDataForContext to take a MeasurementSchedule instead of the
definition id, both the schedule id and definition id can be obtained
and passed to respective implementations.