We do not want invokeServerPluginControl to be transactional
The cassandra server plugin I am working on now exposes a plugin
operation for deploying cassandra nodes via our bundles APIs. One of the
methods called in BundleManagerBean has a txn attribute type of
NOT_SUPPORTED. This resulted in an IllegalStateException. It makes sense
for invokeServerPluginControl to *not* be transactional since we cannot
make any assumptions about the type of plugin operation being invoked.
It may or may not support transactions.