[BZ 838861] Handle empty zip file on java 7
Since an exception is not thrown when closing a ZipOutputStream on an
empty zip file in Java 7, I have added logic to check whether or not the
zip file has any content. sendChangeSetContentToServer is called only if
the content file is non-empty. The logging logic has been updated as
well so that we only log an error message when failing to close the
output stream only when the content file is non-empty. This keeps the
logging clean and consistent across Java 6 and 7.