[BZ 802796] Provide more details for time outs
This commit add more details to the timeout help message.
CreateResourceRunner has also been updated so that it sets the status of
the respone to TIMED_OUT. Previously timed out deployments were getting
flagged as just a FAILURE. The error message in the response now also
contains additional details indicating that the deployment might still
succeed and that the user may want to run a discovery scan to see if it
did.

Conflicts:

	modules/core/plugin-container/src/main/java/org/rhq/core/pc/inventory/CreateResourceRunner.java
	modules/enterprise/gui/coregui/src/main/resources/org/rhq/enterprise/gui/coregui/client/Messages.properties
	modules/enterprise/gui/coregui/src/main/resources/org/rhq/enterprise/gui/coregui/client/Messages_de.properties
	modules/enterprise/gui/coregui/src/main/resources/org/rhq/enterprise/gui/coregui/client/Messages_ja.properties
	modules/enterprise/gui/coregui/src/main/resources/org/rhq/enterprise/gui/coregui/client/Messages_pt.properties