[BZ 848893] Only hide empty groups for availability queries/searches
This regression was introduced during the work for BZ 826493. While
testing the changes for this commit, I came across another, related
issue where an empty compatible group does not show up in the compatible
groups view. It does however show up in the mixed groups view. An empty
mixed group appears in the mixed group view. This issue is already
logged under BZ 708929.