Install rhq cassandra bundle on server start up.
This commit introduces a number of changes in order to support installing the
rhq cassandra bundle on start up. The contents of the build artifact for the
cassandra-bundle module has changed. The JAR file itself is no longer the
bundle. The artifact now consists of two files - the bundle zip file and
a new file cassandra.properties. cassandra.properties will provide deployment
options like the number of nodes in the cluster.

The bundle recipe file, deploy.xml, is now filtered at build time with
cassandra.properties. Attribute values in deploy.xml for things like
the bundle name and bundle version have to be the same as the ones used in
the installBundle method. Filtering with cassandra.properties makes this
possible.

StartupServlet has been updated to invoke CassandraClusterManagerBean.installBundle.
The installBundle method is idempotent. It will install the necessary bundle
components only if they do not already exist in the database. It can be invoked
multiple times without having any additional effect.

Conflicts:

	modules/enterprise/server/jar/pom.xml
	modules/enterprise/server/jar/src/main/java/org/rhq/enterprise/server/core/StartupBean.java
	modules/enterprise/server/jar/src/main/java/org/rhq/enterprise/server/util/LookupUtil.java