Refactoring logic for embedded cluster deployment
This is the initial commit for EmbeddedDeployer. It performs deployment of an
embedded cluster which is the default cluster type for development. An embedded
cluster is one in which all nodes run on localhost and are bound aliases of the
loopback address, e.g., 127.0.0.2, 127.0.0.3, etc.

EmbeddedDeployer does not server APIs to deploy the cluster. More specifically,
it does not use the bundles subsystem. It does however use the Ant-based
bundles APIs to deploy the cassandra bundle. Since it does not use server APIs
which require a running agent to carry out requests, EmbeddedDeployer is
suitable for bootstrapping scenarios.

EmbeddedDeployer looks for the presence of a marker file named .installed in
the cluster directory for installing nodes. If that file exists, it assumes
that the cluster is already is installed and does not do any installation.
If the marker file is not present, anything in the cluster directory is purged
and nodes will be provisioned.