Periodically check that we can connect to at least one cassadra node
This is the initial commit for CassandraClusterHeartBeatJob. It checks that the
RHQ server can connect to at least one cassandra node. If the server cannot
establish a connection to any nodes, the rhq server goes into maintenance mode.
When in maintenance mode the server will be brought back into normal operating
mode when a connection is established.

CassandraClusterManagerBean submits the jobs in its installBundle method which
is called at server start up. EmbeddedDeployer has been updated to return a
list of cluster ndoes. That list is passed into CassandraClusterHeartBeatJob.