Install schema after starting first node
After starting the first node we wait until we can establish a connection to it
in order to verify that it is running. If we cannot connect after 10 attempts,
an exception is thrown. Once a connection is established, we execute
cassandra-cli on dbsetup.script.

There is still a lot more work to do here like better error handling, logging,
etc.