adding work around for snappy-java library with Java 7
Cassandra uses snappy-java, a library for the snappy compressor. It uses a
native library. That library fails to load in Cassandra when running on Mac
OS X with Java 7 due to https://github.com/xerial/snappy-java/issues/6. This
issue is fixed in newer versions of snappy-java. This commit adds a profile
that is activated when running on Mac OS X and that upgrades the version
of snappy-version in the cassandra bundle so that the snappy compressor can
still be used during development/testing.