Initial commit for CassandraClusterManager
CassandraClusterManager (which maybe should be renamed to something along the
lines of TestClusterManager or something to indicate its for testing)
encapsulates cluster deployment and shutdown. It it used as a TestNG listener
in conjuntion with the new @DeployCluster and @Shutdown annotations. This
approach offers more flebility than providing a common base class with the
same functionality. Attributes will be added to the annotations for configuring
cluster deployment like the number of nodes to install and run.

The CassandraClusterManager.shutdownCluster method uses JNA to kill cluster
nodes. Initially I was going to use sigar, but the required functionality is
not available on all development platforms.