Adding a ping method to ClusterInitService
The ping method is similar to waitForClusterToStart except that it only goes
through all of the hosts at most once. If connections are made to to the
specified number of hosts the ping method immediately returns true whereas
waitForClusterToStart will block until the specified number of connections
is made. CassandraClusterHeartBeatJob has been refactored to use the new ping
method.