[BZ 869015] Add more detail logging when handling inventory report with stale types
This commit contains a few changes:

* Add logging in DiscoveryBossBean when inventory report contains deleted types
There is detailed log statements that include both the deleted resource types
as well as the agent sending the report.

* Remove ERROR log statement from DiscoveryServerServiceImpl
No need to log the exception since it is not an error. We just want to throw
the exception back to the agent so it knows to resync its plugins.

* Revert ResourceType.toString
Not sure why but in commit bb2b881045 was changed. The version prior to that
commit was already concise yet detailed and specific. I have reverted back
to that version.