update start up logic to first look at system properties for cluster configuration
Since cluster configuration properties are now in rhq-server.properties they are
made available as system properties. Now at start up we first look cluster
config property values in system properties and then fall back to the values in
cassandra.properties for defaults if system properties are not defined. Logic
has also been added to only do the unmanaged deployment if it is an embedded
cluster.