upgrading cassandra bundle to cassandra 1.2.0-beta1
This commit includes some substantial changes. Previously the cassandra distro
was stored in the module source tree. The distro has been removed from the
source tree and it is now downloaded during the build if it is not already on
disk. The cassandra distro is in the central maven repo, but it is not available
as a maven artifact; consequently, we are downloading it in the ant packaging
script.

There are a number of changes and new features in cassandra 1.2.0. Because of
the new virtual nodes support, we no longer need to specify an initial token.
See http://wiki.apache.org/cassandra/VirtualNodes/Balance and cassandra.yaml
for details.

In previous versions of the bundle recipe we set the auto_bootstrap property to
false. That property is no longer in cassandra.yaml. I think it is still
recognized; however, in org.apache.cassandra.config.Config, the auto_boostrap
field is hard coded to true.

I am not sure what the ring delay defaulted to in prior releases but now it
defaults to 30 seconds. This results in a 30 second pause at start up which
is not ideal for automated tests. I have udpated cassandra-env.sh so that we
can override the ring delay for tests via the rhq.cassandra.ring.delay
property.