refactoring ant packaging scripts
A couple changes have been made to the embedded ant scripts. First, the
cassandra distro is now downloaded from the central maven repo and stored in
the local maven repo. This makes things cleaner and helps prevents having to
repeatedly downlod the distro as was the case when it was being stored in the
target directory.

The second change is to accomodate the Mac OS X work around for the snappy java
library. The distro is now unpacked in the generate-resources phase and the
snappy-mac-workaround phase now executes its ant script in the process-resources
phase. This allows for the snappy-java library replacement to occur before the
bundle zip file is generated.