Moving logic out of DeploymentOptions constructor and into new load method
The default constructor of DeploymentOptions initialized its properties first
looking at system properties then at values in cassandra.properties. This was
kind of awkward for tests as they had to populate a Properties object and pass
that into the DeploymentOptions(Properties) constructor. The logic in the
default constructor has been moved to a new load method. This makes it easier
for tests to simply call setter methods on properties then invoke the load
method to get default values for other, remaining properties.