(cql port) first cut at aggregating raw data
This is the first of many refactorings for aggregating using CQL instead of
hector/thrift. This commit updates the aggregateRawData method and some
corresponding test code in MetricsServerTest.

This is also the initial commit for AggregatedNumericMetric and ColumnMetadata.
Right now these classes are just used in test code for verification. I
considered using some of the classes in core/domain like
MeasurementDataNumeric1H but chose not to for a few reasons.

1) There are separate, duplicate classes for 1H, 6H, and 1D. I suspect that
mulitple classes were introduced to support hibernate/jpa mappings to different
tables. A single class is sufficient.

2) The classes are cluttered with JPA mappings that are not applicable any
more. On the one hand they do not hurt anything, but on the other hand they
can cause confusion.

3) Those classes have a schedule property which is of type MeasurementSchedule.
Stuff is cassandra is denormalized and will only store a schedule id. There is
no need for the cassandra version of the corresponding domain objects to
maintain a reference to the full, complete MeasurementSchedule. The id is
sufficient and also makes testing easier.