adding methods to MetricsDAO to insert and retrieve aggregates
Adding the methods insertAggregates and findAggregateMetrics to MetricsDAO.
This is also the initial commit for AggregateMetricMapper which is used by
findAggregateMetrics.

MetricsServer.aggregateRawData has been further refactored so that it is no
longer direct JDBC calls and is instead using the API provided by MetricsDAO.
More changes are forthcoming to address a number of things including:

* setting TTL on metrics getting inserted
* defining the consistency level for each read/write operation
* handling exceptions when/where it makes sense

Right now, every checked SQLException is wrapped in an unchecked CQLException
and nothing more is done. Unlike with the RDBMS, there will be times we want to
recover from such exceptions. For instance, if an exception is thrown because
the cassandra node is down, then we probably want to retry the operation with
another node.