(cql port) adding back support for producing 6 hr aggregates
MetricsServer.calculateAggregates has been refactored to use MetricsDAO in
place of the Hector APIs previously used. This commit also fixes a subtle bug
in test code for getting hour zero of the current day.

There are still some problems/inconsistencies with how timestamps for metrics
index entries are generated. That needs to be addressed in a subsequent commit.