refactoring code to update metrics index
Prior to this commit some of the logic for updating the metrics index was in
MetricsDAO while the rest of it was in MetricsServer. MetricsDAO is intended to
be singularly focused around data access. All of the logic for updating for
updating the metrics index is now in MetricsServer. This involves determining
which row to update and calculating the appropriate time slices.

For example, if we insert 1hr data at 05 hours (i.e., 5:00 am), the six hour
row in the index table should be updated with a timestamp of 00 hours
(i.e., 12:00 am) since 5:00 am falls into the 00 - 06 time slice.