adding option to skip cluster shutdown during test runs
The @ShutdownCluster annotation causes each node in the cluster to be shut
down. This type of clean up is particularly useful in a continuous integration
environment. But in a dev environment when running tests repeatedly prior to a
commit for example, it gets to be cumbersome to redeploy/restart the cluster
for each test run.

CassandraClusterManager now checks a system property named
rhq.cassandra.cluster.skip-shutdown to decide whether or not cluster shut down
should be initiated when the @ShutdownCluster annotation is present.

The server-metrics build has been updated to pass in the system property when
running tests from the command line. You can easily disable shut down as
follows,

mvn test -Dtest=MetricsDAOTest -DskipClusterShutdown=true

If there are multiple test/configuration methods with the @ShutdownCluster
annotation, then -DskipClusterShutdown=true will affect all of them.