CassandraClusterManager should not automatically purge cluster directory
CassandraClusterManager was purging the cluster deployment directory prior to
running BootstrapDeployer which means the cluster would always get deployed
or redeployed. This would be problematic if nodes are still running. In a
future commit, I may add an attribute to @DeployCluster that will trigger
purging the cluster directory. Prior to purging though, we need to make sure
all nodes are shut down.