purge one hour metrics index after calculating raw aggregates
After metric data is aggregated or compressed, we need to delete the
corresponding metrics index entries; otherwise, we would wind up needlessly
recalculating those aggregates again. In the current implementation, the whole
row in the index table is deleted. We can safely delete the entire row for two
reasons. 1) When aggregates are calculated for a particular metrics table, the
entire row in the metrics index table is read. We go through the whole index
for the table being compressed. 2) There are no other writers to the index
table; so, we do not have to worry about losing index entries yet to be
processed. In the future, if this aggregation work flow because multi threaded
or multi process then we will need to delete by schedule id and timestmap.