adding support for setting column meta data when inserting raw metrics
Each column in cassandra has a time to live (TTL) and timestamp attributes. If
the TTL is not set, then the column will not expire unless explicitly deleted.
If the TTL is set, Cassandra will delete the column after the specified number
of seconds has passed. The timestamp attribute is basically the insertion time.
It is used for conflict resolution between replicas. The TTL for raw metrics is
set to seven days.