[BZ 845700] Do not use SQL LIKE operator when checking if plugin is ready for purge
The query used in PluginManagerBean.isReadyForPurge used the LIKE operator to
make sure all resource types have been deleted. This is problematic with the
jboss-as (JBoss 4) plugin. Its name is JBossAS and the jboss-as-5 plugin name
is JBossAS5. When using the LIKE operator on the plugin name, both plugin names
match which prevents the purge from happening. The LIKE operator is no longer
used to prevent this problem. A test has been added to PluginManagerBeanTest
for this scenario.