[BZ 857189]  change transaction boundaries when installing resource types
Previously a new transaction was started for
PluginManagerBean.registerPluginTypes. All resource types and associated meta
data are persisted to the database during this method. For a plugin like the
AS 7 plugin which has a var large number of resource types (406 to be exact),
this could result in transaction time outs which in turn would lead to the
plugin not to be fully installed.

This commit changes the transaction boundaries such that a transaction is no
longer started for the registerPluginTypes method. Instead a new, separate
transaction is created for each individual resource type that is being
installed/updated.