[BZ 870561] Fix data points query to use minvalue and maxvalue columns
The data points query previously only used the value or avg of metrics for the
hi and low values for each data point returned. If the query is against tables
other than raw, the minvalue and maxvalue columns are used.

This is also the initial commit for MeasurementDataManagerBeanTest. There are
two tests so far to verify the correct results for the data points query when
querying against the raw tables and then against the 1hr table.