Initial commit for SchemaManager
SchemaManager handles creating and installing the schema. Most of the work is
delegated to Liquibase. Previously the work was done in BootstrapDeployer
which lives in the cassandra-common module. We only need to run Liquibase
in the installer and deploying it in the EAR causes problems with CDI due the
JSF libraries conflict in portal-war. The logic has been pulled out of
BootstrapDeployer and is now only used in the installer.