first cut at integrating cassandra into new installer
This is the first of what will be several commits for integrating with the new
installer. The installer build has been updated so that we now deploy all of
the cassandra-related dependencies in the installer module. For now version
strings are hard coded because I am hoping that a number of these dependencies
will go away once I open a pull request against the datastax driver.

This commit also includes an initial update to the installer code to call
SchemaManager to create the schema (if necessary) and then apply any schema
updates via liquibase.