fixing a few issues with test failures
Test clean was failing because we were not purging the resource from the db.
The uuid constant in the test class did not match the one actually used.
When setting the uuid, Resource pads it out to 36 characters.

The other issue involves dealing with decimal precision. I have added a
function to use up to 4 decimal places.