[BZ 880795] - The upgraded Augeas library no longer contains the file "libaugeas.so", which causes the default native library lookup mechanisms to fail. Standard ld should be able to find the .so.x.y files, too, but the presence of java.library.path system property seems to be causing this not to kick in during the library loading as done by the JVM.
To work around this problem, we need to instruct JNA to have a "backup
plan" in case the standard lookup mechanisms fail. This can be done using
the "jna.platform.library.path" system property, which instructs JNA to use
that path to look up libraries if everything else fails (and do the .so.x.y
resolution on its own).

We set this variable to the same value as we override LD_LIBRARY_PATH to
so that JNA library loading and native (as in system-wide) library loading
works the same for the agent process. This way we can load the JNA
libraries using the .so.x.y resolution the same way as the system resolves
the native libraries using stadard ld mechanisms in spite of there being
the java.library.path variable set (which is needed by SIGAR).
(cherry picked from commit e437f5771c4b79ce570ad24cb3c545078d9a944a)