major refactoring of the former cassandra-common module
The cassandra-common module and cassandra-common-itests module were evolving
into a few different things so this was a good time to reorganzie things.

We now have rhq/common/cassandra-ccm which is a parent pom. Under it we have
cassandra-ccm-core which contains most of the stuff that was previously in
cassandra-common.

We also have cassandra-ccm-cli which contains the new CLI
class and this maven module produces the JBoss AS module that is run by the
rhq-ccm script.

Lastly we have cassandra-ccm-testng which contains the classes that were in
cassandra-common-itests.