[Bug 886850 - 4.6 on as7: different rollback semantic for Exceptions] It seems AS 4.2.3 did not wrap RuntimeExceptions as EJBExceptions as per the spec. Neither did it rollback a supported transaction. Furthermore, RHQ code was written assuming the bad semantics were correct. AS7 corrects the handling and therefore the RHQ code is now in error in certain places.
AS 4.2.3 basically treated the RuntimeException as an ApplicationException. This
commit tries to get sime consistency around our Exception classes that should
be treated this way.  It adds the @ApplicationException annotation to several
classes, and a few others are changed from extending Exception to extending
RuntimeException as well as then being annotated.

A test has been added that ensures handling of the relevant exception classes
works as expected, as an ApplicationException that is not wrapped and does
not rollback a supported transaction.

Also:
- revert changes to UserHandlerBean to work around the problem reported in this BZ
- fix catch statement in ResourceOperationJob to catch thre ApplicationException, not the wrapped EJBException
- fix issue in OperationManagerBean to throw the proper ApplicationException