major refactoring of BootstrapDeployer
The functionality of BootstrapDeployer has been streamlined so that it focuses
on node deployment. The work involved with calculating various config
properites for an embedded deployment is now handled by CassandraClusterManager
in the createCluster method. This makes more sense because BootstrapDeployer
is used for both dev and production deployment, but embedded deployments are
only for dev/testing. The createCluster method in CCM is specific to embedded
deployments.

We are now also utilizing the @BundleProperty annotation introduced in a prior
commit. Properties defined in cassandra.properties or as system properties are
merged into a DeploymentOptions object. A number of the properties in that
object are used to populate a Properties object that is passed to the bundle
deployment. Keeping that mapping of DeploymentOptions --> Properties was
tedious and error prone.

The mapping has largely been automated. Any property in DeploymentOptions that
has the @BundleProperty annotation is added to the Properties object that is
ultimately passed to ant.